create new component file ProductListItem.tsx and refactor index.tsx
