+++ src/app/(tabs)/index.tsx
@@ -1,39 +1,13 @@
-import Colors from '@/src/constants/Colors';
-import { StyleSheet, Text, View, Image } from 'react-native';
+import { View } from 'react-native';
 import products from '@/assets/data/products';
+import ProductListItem from '@/src/components/ProductListItem';
 
-const product = products[1];
-
-export default function TabOneScreen() {
+export default function MenuScreen() {
   return (
-    <View style={styles.container}>
-      <Image source={{ uri: product.image }} style={styles.image}/>
-      <Text style={styles.title}>{product.name}</Text>
-      <Text style={styles.price}>${product.price}</Text>
+    <View>
+      <ProductListItem product={products[0]}/>
+      <ProductListItem product={products[1]}/>
     </View>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: 'white',
-    padding: 10,
-    borderRadius: 20,
-  },
-  title: {
-    fontSize: 18,
-    fontWeight: '600', 
-    marginVertical: 10,
-
-  },
-  price: {
-    color: Colors.light.tint,
-    fontSize: 18,
-    fontWeight: '600',
-  },
-  image:{
-    width: '100%',
-    aspectRatio: 1,
-  }
-
-});
