+++ src/components/ProductListItem.tsx
@@ -1,0 +1,42 @@
+import Colors from '@/src/constants/Colors';
+import { StyleSheet, Text, View, Image } from 'react-native';
+import products from '@/assets/data/products';
+
+
+const ProductListItem =({ product }) => {
+  return (
+    <View style={styles.container}>
+      <Image source={{ uri: product.image }} style={styles.image}/>
+      <Text style={styles.title}>{product.name}</Text>
+      <Text style={styles.price}>${product.price}</Text>
+    </View>
+  );
+};
+
+export default ProductListItem
+
+
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: 'white',
+    padding: 10,
+    borderRadius: 20,
+  },
+  title: {
+    fontSize: 18,
+    fontWeight: '600', 
+    marginVertical: 10,
+
+  },
+  price: {
+    color: Colors.light.tint,
+    fontSize: 18,
+    fontWeight: '600',
+  },
+  image:{
+    width: '100%',
+    aspectRatio: 1,
+  }
+
+});
